% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{APAtable_ROPE}
\alias{APAtable_ROPE}
\title{Produce an APA formatted ROPE test table}
\usage{
APAtable_ROPE(rstan_model)
}
\arguments{
\item{stan_glm_model}{An stan_glm model object.}
}
\description{
Accepts a Bayesian Logistic Regression rstanarm model and returns a perfectly formatted APA results table with flextable, ready to be exported to MS Word.
}
