% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{bundle_paths}
\alias{bundle_paths}
\title{Create a list of file info for quick writing}
\usage{
bundle_paths(df_list, output_location, file_type)
}
\arguments{
\item{df_list}{A named list of data frames}

\item{output_location}{The desired folder location where the files are to be saved. Uses regular format, NOT the \code{here} package}

\item{file_type}{A string that contains the file type}
}
\description{
Grab and store a list of file names, extensions, and output locations to be sent to purrr for writing many files. This function creates the second of the two lists needed for exporting many files at once with \code{walk2}
}
