% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching_functions.R
\name{calc.variability.table}
\alias{calc.variability.table}
\title{Calculate and show variability table}
\usage{
calc.variability.table(...)
}
\arguments{
\item{...}{A list of numbers}
}
\description{
This function accepts a list of numbers and returns (a) the Sum of Squares, (b) the sample variance and standard deviation, (c) the sample mean, and (d) a complete table showing every step of each calculation for (a) and (b).
}
\examples{
calc.variability.table(1,7,2,8,4,9,3,3,6,2)

}
