% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{convert_spss}
\alias{convert_spss}
\title{Convert an SPSS .sav file to .csv}
\usage{
convert_spss(sav_file, csv_name)
}
\arguments{
\item{sav_file}{The name of the SPSS .sav file, including the .sav extension.}

\item{csv_name}{The name you want to give to the converted .csv file, including the .csv extension. Can be the same as or different from the original .sav name.}
}
\description{
Does what the name implies; converts an SPSS .sav file in the working directory to .csv file format. Please note that the argument names must be in quotes, and must include the file extensions.
}
\examples{
convert_spss(sav_file="survey.sav", csv_name="converted_file.csv")
}
