% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{count_duplicates}
\alias{count_duplicates}
\title{Count duplicates}
\usage{
count_duplicates(df, col, character_data = TRUE)
}
\arguments{
\item{df}{Your data set}

\item{col}{The specific column within the data set to look in (e.g., a column of participant's names or emails). UNQUOTED!!!}

\item{character_data}{Indicate with TRUE or FALSE whether or not the supplied column is a vector of character data or not.}
}
\description{
This function searches though a data frame for duplicate entries to ensure that extra responses don't mess up your data analysis. The function returns a list that shows a table with the observed counts of responses and the total count of the extra responses (assuming that each person is supposed to have one response each)
}
\examples{
count_duplicates(mtcars,cyl, character_data=FALSE)
}
