% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{count_missing}
\alias{count_missing}
\title{Count the percentage of missing data}
\usage{
count_missing(df)
}
\arguments{
\item{df}{A data frame (or a subset of a data frame) to look for missing data in.}
}
\description{
Quickly tabulates the number of NA's in a data set.
}
\examples{
mtcars \%>\% count_missing()
}
