% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching_functions.R
\name{curve.grades}
\alias{curve.grades}
\title{Curve Quiz Grades}
\usage{
curve.grades(quiz_df, curve.points)
}
\arguments{
\item{quiz_df}{A Canvas quiz data frame}
}
\description{
I'm feeling to lazy right now to right this description.
}
