% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{describe_odds}
\alias{describe_odds}
\title{Grab and display odds ratio's for parameters in a Bayesian Logistic Regression}
\usage{
describe_odds(stan_glm_model)
}
\arguments{
\item{stan_glm_model}{An rstanarm model object (Binomial Logistic Regression models only).}
}
\value{
A nice summary table.
}
\description{
Combination of bayestestR::describe_posterior and effectsize::interpret_oddsratio that gives a summary table of parameters and their respective odds ratios, for rstan BLR models
}
