% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{describe_posterior_fancy}
\alias{describe_posterior_fancy}
\title{Fancier model summary}
\usage{
describe_posterior_fancy(model)
}
\arguments{
\item{model}{An rstanarm model object. Since this command adds an Odds Ratio column, it is made for Binomial Logistic Regression models.}
}
\value{
The posterior summary.
}
\description{
This function uses bayestestR::describe_posterior to generate a summary of a Bayesian rstan model. It adds an Odds Ratio column for Logistic Regressions, and neatens some columns up for nicier presentation.
}
