% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{drop_dupes}
\alias{drop_dupes}
\title{Remove duplicate strings}
\usage{
drop_dupes(df, x)
}
\arguments{
\item{df}{A data frame}

\item{x}{The column to check for duplicates. Must be of class 'character'.}
}
\description{
This function is designed to find and remove duplicate survey entries from SONA studies. You provide a data set and a column in which to check for duplicates (e.g., participant's names or email addresses), and this function removes all rows from the data that appear more than once. For example, if "Renee Johnson" completed the study three times, this function will keep only her first response and delete the later two entries.
}
\examples{
my_survey = my_survey \%>\% drop_dupes(name)
}
