% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{drop_junk_responses}
\alias{drop_junk_responses}
\title{Remove test runs and spam from Qualtrics data files}
\usage{
drop_junk_responses(df)
}
\arguments{
\item{df}{A Qualtrics survey import that has the "status" column.}
}
\description{
In a Qualtrics survey, the "status" column contains the response type code. Responses will be flagged as 1 for real participants, or other numbers for spam responses or test runs. This command removes all test runs and spam responses (i.e., responses not equal to 1) from the data set, based on their coding. It then drops the status column when finished.
}
