% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{drop_Qualtrics_JunkCols}
\alias{drop_Qualtrics_JunkCols}
\title{Remove seldom-used variables from a fresh Qualtrics survey import}
\usage{
drop_Qualtrics_JunkCols(df)
}
\arguments{
\item{df}{An imported Qualtrics survey}
}
\description{
Searches a Qualtrics import and removes the columns: "progress", "finished", "distribution_channel", "user_language", and "recorded_date".
}
\examples{
drop_Qualtrics_JunkCols(df)
}
