% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Find duplicates}
\usage{
find_duplicates(df, col)
}
\arguments{
\item{df}{Your data set}

\item{col}{The specific column within the data set to look in (e.g., a column of participant's names or emails)}
}
\description{
This function searches though a data frame for duplicate entries to ensure that extra responses don't mess up your data analysis. The function returns a list that shows a table with the observed counts of responses and the total count of the extra responses (assuming that each person is supposed to have one response each)
}
\examples{
find_duplicates(mtcars,cyl)
}
