% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{gen_id}
\alias{gen_id}
\title{Generate random ID tags}
\usage{
gen_id(n)
}
\arguments{
\item{n}{The number of names to generate. Set equal to the desired number. If inserting the generated list of names into a table, set to the same number of rows as the table.}
}
\description{
Generate random ID's for participants to ensure confidentiality or any other purpose. Each random ID consists of a first name followed by a random number.
}
\examples{
tibble(name=gen_id(n=3), fav_color=c("red", "blue", "green"))
}
\seealso{
The list of names used in this function were randomly drawn from the US Baby Names data set, downloadable here: https://www.kaggle.com/kaggle/us-baby-names
}
