% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{how_much_coffee}
\alias{how_much_coffee}
\title{Calculate Coffee Strength}
\usage{
how_much_coffee(desired_strength, grams_water, use_ratios)
}
\arguments{
\item{desired_strength}{The strength you want the coffee to be. Can be either in grams per liter, or a strength ratio}

\item{grams_water}{The amount of water you plan to use for the brew.}

\item{use_ratios}{a TRUE or FALSE toggle to tell the formula what your input will be for the desired strength: set to FALSE if using grams per liter, or TRUE if using typical ratios.}
}
\description{
Quickly calculates how many grams of coffee you need for a desired amount of water, to reach a given dosage strength in grams per Liter
}
