% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching_functions.R
\name{import.quiz}
\alias{import.quiz}
\title{Quick Import Canvas Quiz}
\usage{
import.quiz(quiz_file, max_points)
}
\arguments{
\item{quiz_file}{A Canvas quiz .csv export (student analysis version only, not item analysis)}

\item{max_points}{The maximum total number of points possible on the quiz. If someone got this score, they would have a 100\%.}
}
\description{
Import the results of a Canvas quiz
}
