% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{load_bayes_suite}
\alias{load_bayes_suite}
\title{Quickly load data analysis packages}
\usage{
load_bayes_suite(x)
}
\description{
Includes the whole easystats suite, flextable, bayestestR, and rstanarm.
}
