% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{loo_compare_tidy}
\alias{loo_compare_tidy}
\title{Display LOOIC and WAIC info for multiple models}
\usage{
loo_compare_tidy(...)
}
\arguments{
\item{list}{Multiple brms fit objects to compare}
}
\description{
Quickly sum up and neatly present Info Criterion with a quick command
}
