% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{openGraph}
\alias{openGraph}
\title{Open a graphics window}
\usage{
openGraph(width = 7, height = 7, mag = 1, ...)
}
\arguments{
\item{width}{The width of the window}

\item{height}{The height of the window}
}
\description{
Opens the Plots viewer as a separate window, at the dimensions specified by the user.
}
\examples{
openGraph(width= 5, height= 6)
}
