% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{prime_r}
\alias{prime_r}
\title{Load essential packages for analyzing data}
\usage{
prime_r()
}
\description{
This function sets up R for data analysis by loading the following packages: The whole easystats suite, rstanarm, flextable, psych, broom, and broom.mixed.
}
