% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{read_all}
\alias{read_all}
\title{Read and import all files in working directory}
\usage{
read_all(path, extension)
}
\arguments{
\item{path}{The path in the working directory where the files are located. Best to specify with slashes rather than using the here package.}

\item{extension}{The file extension of the files you want to import. Can be either ".xlsx" for Excel or ".csv" for CSV}
}
\description{
Quickly import multiple files at once with this command. Note that this reads all files in and stores them as separate data frames.
}
