% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{read_spss}
\alias{read_spss}
\title{Read in data from an SPSS file}
\usage{
read_spss(file)
}
\arguments{
\item{file}{An SPSS .sav file. Note that the full file name and file extension should be included, and the whole thing quoted.}
}
\description{
Read data in from any SPSS file and store it as a tidy tibble.
}
\examples{
read_spss("survey.sav")
}
