% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{save_table}
\alias{save_table}
\title{Save a table for exporting}
\usage{
save_table(flextable_object, file_path, file_name)
}
\arguments{
\item{flextable_object}{The flextable you want to export}

\item{file_path}{The path of the file directory you want to save it in; i.e., where the file should be saved.}

\item{file_name}{The name you want to give the file. Must also include the file extension (".docx") at the end}
}
\description{
Shortcut function that combines to commands into one. Makes it super easy to export a flextable object to MS Word.
}
