% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{saveGraph}
\alias{saveGraph}
\title{Save a graph in the Plots viewer/graphics window}
\usage{
saveGraph(file = "saveGraphOutput", type = "pdf", ...)
}
\arguments{
\item{file}{The desired file name for the output}

\item{height}{Desired file type. Opetions are: PDF, jpg, or eps.}
}
\description{
Save a graph in the Plots viewer/graphics window
}
\examples{
saveGraph(file= "a_pretty_graph", type="jpg")
}
