% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{splice_vars}
\alias{splice_vars}
\title{Splice together data sets}
\usage{
splice_vars(numeric_df, text_df)
}
\arguments{
\item{numeric_df}{A qualtrics .csv file export.}

\item{text_df}{A Qualtrics .csv file import}
}
\description{
Description.
}
\examples{
data= read_Qualtrics("survey.csv")
}
