% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{style_ggplot}
\alias{style_ggplot}
\title{Quick styling options}
\usage{
style_ggplot(gg_graph, center_title, remove_gridlines)
}
\arguments{
\item{center_title}{Centers the title on a ggplot figure. Two-option toggle that can be set to TRUE or FALSE.}

\item{remove_gridlines}{Removes the grid lines on a ggplot figure. Two-option toggle that can be set to TRUE or FALSE.}
}
\description{
Shortcut function that combines to commands into one: Center a title and remove grid lines on a ggplot figure.
}
