% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{style_table}
\alias{style_table}
\title{Apply some standard APA format options to a flextable}
\usage{
style_table(flextable_object, table_title, table_number, add_note)
}
\arguments{
\item{flextable_object}{The flextable to be modified}

\item{table_title}{The title you want to add to the table}

\item{table_number}{The table number}
}
\description{
Quickly set the font size to 11; the font style to Times New Roman; add two header lines with the table number and title; and auto fit all columns.
}
