% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching_functions.R
\name{summarize_grades}
\alias{summarize_grades}
\title{Quick Summary of Quiz Grades}
\usage{
summarize_grades(quiz_df)
}
\arguments{
\item{quiz_df}{A Canvas quiz data frame imported with the import.quiz function}
}
\description{
This accepts a quiz data frame and returns a summary table of both the un-curved and curved quiz grades
}
