% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{summarize_pleas}
\alias{summarize_pleas}
\title{Summarize plea acceptance across conditions}
\usage{
summarize_pleas(data, dv, ...)
}
\arguments{
\item{data}{Your data set}

\item{dv}{The specific column within the data set to look in (e.g., a column of participant's names or emails)}

\item{...}{A list of your IV variable names}
}
\description{
Build a descriptive summary table of guilty pleas by experimental condition. Table contains the total number of participants per condition, the total number of guilty pleas per condition (and the percentage of n), and the condition names and values.
}
\examples{
summarize_pleas(data=sonadata, dv=accept_reject, discount_level, pts_level, evidence_strength)
}
