% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{table_apa_BayesLogReg}
\alias{table_apa_BayesLogReg}
\title{Produce APA formatted Bayesian model summary table}
\usage{
table_apa_BayesLogReg(stan_glm_model)
}
\arguments{
\item{stan_glm_model}{An stan_glm model object.}
}
\description{
Accepts a Bayesian Logistic Regression rstanarm model and returns a perfectly formatted APA results table with flextable, ready to be exported to MS Word.
}
