% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{table_apa_BFmodels}
\alias{table_apa_BFmodels}
\title{Produce an APA formatted Bayes factor models table}
\usage{
table_apa_BFmodels(...)
}
\description{
Accepts as many rstan models as you want, and returns an APA results table with flextable, ready to be exported to MS Word.
}
