% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{tidy_bf_models}
\alias{tidy_bf_models}
\title{Produce an APA formatted Bayes factor models table}
\usage{
tidy_bf_models(bayestestR_obj)
}
\description{
Turn a single BF model object into a tidy tibble with interpretations, thanks to bayestestR
}
