% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{tidy_date}
\alias{tidy_date}
\title{Quickly tidy dates}
\usage{
tidy_date(col, includes_timestamp = TRUE)
}
\arguments{
\item{col}{The date column you want to change.}
}
\description{
Combines two formating functions in one to make working with dates less terrible. The first thing the function does is take a character vector of the format MM/DD/YYYY and tell R to change it to the more stat-software-friendly-format YYYY-MM-DD. It then converts this reformatted character string into a date object with the lubridate package.
}
