% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{tidy_info_criterion}
\alias{tidy_info_criterion}
\title{Display LOOIC and WAIC info for multiple models}
\usage{
tidy_info_criterion(...)
}
\arguments{
\item{list}{Multiple brms fit objects}
}
\description{
Quickly sum up and neatly present Info Criterion with a quick command
}
