% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{drop_dupes}
\alias{drop_dupes}
\title{Drop duplicate responses from the data set}
\usage{
drop_dupes(df, col)
}
\arguments{
\item{df}{the data frame}

\item{col}{the UNQUOTED column to drop duplicate entries from. MUST BE a character string (otherwise just use the distinct verb from dplyr).}
}
\description{
Drop duplicate responses from the data set
}
