% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teaching_functions.R
\name{import.quiz}
\alias{import.quiz}
\title{Quick Import Canvas Quiz}
\usage{
import.quiz(quiz_file)
}
\arguments{
\item{quiz_file}{A Canvas quiz .csv export}
}
\description{
I'm feeling to lazy right now to right this description.
}
