% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{mark_outliers}
\alias{mark_outliers}
\title{Find outliers in a sample of data}
\usage{
mark_outliers(df, col, time_data = TRUE)
}
\arguments{
\item{df}{The supplied data set}

\item{col}{The column to examine for outliers. If time_data is set to TRUE, \eqn{a + b}}
}
\description{
Search through an indicated column for scores that are outliers, and marks them with a 0/1 coding. Depending on the formula used....
}
