% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{read_Qualtrics}
\alias{read_Qualtrics}
\title{Quick import Qualtrics files}
\usage{
read_Qualtrics(file)
}
\arguments{
\item{file}{A qualtrics .csv file export.}
}
\description{
A quick-import option for Qualtrics surveys that performs a number of functions. It reads in a .csv file; cleans the column names; removes the extra two rows that Qualtrics includes underneath the header rows; reformats the date columns and removes their timestamps; drops junk columns; and drops test runs and spam responses from the data.
}
\examples{
data= read_Qualtrics("survey.csv")
}
