% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{splice_text}
\alias{splice_text}
\title{Splice together data sets}
\usage{
splice_text(numeric_df, text_df)
}
\arguments{
\item{numeric_df}{A qualtrics .csv file export containing the NUMERIC versions of responses}

\item{text_df}{A Qualtrics .csv file import containing the TEXT versions of responses}
}
\description{
Description.
}
\examples{
data= read_Qualtrics("survey.csv")
}
